requirements.txt
astroid==1.6.0
boto==2.48.0
cssselect==1.0.3
cssutils==1.0.2
dj-database-url==0.4.2
Django==1.11
django-appconf==1.0.1
django-compressor==2.1
django-pagedown==1.0.4
django-ses-backend==0.1.1
djmail==0.11.0
flake8==3.5.0
futures==3.0.3
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
pbr==3.1.1
premailer==2.9.6
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.1
pytz==2017.3
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
